refactoring and change init routers
